refactor(tests): move audit-round regression files to their owning components (#2000)
Move 6 regression tests from test_audit_fixes.py and test_audit_fixes_round2.py
to their semantic owners: observation_pair_failures -> test_benchmark_contracts,
usage stats -> test_heldout_runner, trial_status + observation_failures ->
test_observation_artifact_binding, compare_evidence -> test_observation_comparison.
Both audit-round files deleted.
